reorganize task order routes
