+++ templates/portfolios/header.html
@@ -67,8 +67,8 @@
     {{ Link(
       icon='dollar-sign',
       text='navigation.portfolio_navigation.breadcrumbs.funding' | translate,
-      url=url_for("portfolios.portfolio_funding", portfolio_id=portfolio.id),
-      active=request.url_rule.endpoint == "portfolios.portfolio_funding",
+      url=url_for("task_orders.portfolio_funding", portfolio_id=portfolio.id),
+      active=request.url_rule.endpoint == "task_orders.portfolio_funding",
     ) }}
     {% if user_can(permissions.VIEW_PORTFOLIO_ADMIN) %}
       {{ Link(
