+++ templates/portfolios/reports/index.html
@@ -76,7 +76,7 @@
                       -
                   {% endif %}
                 </div>
-                <a href='{{ url_for("portfolios.view_task_order", portfolio_id=portfolio.id, task_order_id=task_order.id) }}' class='icon-link'>
+                <a href='{{ url_for("task_orders.view_task_order", task_order_id=task_order.id) }}' class='icon-link'>
                   {{ Icon('cog') }}
                   Manage Task Order
                 </a>
