+++ templates/portfolios/task_orders/index.html
@@ -8,7 +8,7 @@
 {% block portfolio_content %}
 
 {% macro ViewLink(task_order) %}
-  <a href="{{ url_for('portfolios.view_task_order', portfolio_id=portfolio.id, task_order_id=task_order.id) }}" class="icon-link view-task-order-link">
+  <a href="{{ url_for('task_orders.view_task_order', task_order_id=task_order.id) }}" class="icon-link view-task-order-link">
     <span>View</span>
     {{ Icon("caret_right", classes="icon--tiny") }}
   </a>
