+++ templates/portfolios/task_orders/review.html
@@ -19,7 +19,7 @@
     {% include "fragments/flash.html" %}
 
     {% block form_action %}
-    <form method='POST' action="{{ url_for('portfolios.submit_ko_review', portfolio_id=portfolio.id, task_order_id=task_order.id, form=form) }}" autocomplete="off" enctype="multipart/form-data">
+    <form method='POST' action="{{ url_for('task_orders.submit_ko_review', task_order_id=task_order.id, form=form) }}" autocomplete="off" enctype="multipart/form-data">
     {% endblock %}
 
     {{ form.csrf_token }}
